refactor: split out the upload and download logic into builtins.Cat() and builtins.Redirect()
This is entirely refactoring, there should be neither features nor semantic changes.
